requirements.txt
scrapy==2.11.0
PyYAML>=6.0